requirements.txt
python-telegram-bot==12.0.0b1
google-api-core==1.13.0
redis==3.3.8
vk_api==11.5.0